extract structs from tokenizer_manager.rs to its own files
