+++ rust/sglang-server/src/tokenizer_manager/detokenizer.rs
@@ -28,7 +28,7 @@ use crate::message::egress::{ChunkEvent, EgressItem, EgressSink, SinkError};
 use crate::message::finish_reason::Matched;
 use crate::message::ids::Rid;
 use crate::message::types::TokenIds;
-use crate::tokenizer_manager::AbortSource;
+use crate::tokenizer_manager::wiring::AbortSource;
 use crate::utils::runtime::Runnable;
 use crate::utils::{
     error::Error,
